fix msvc performance bug
msvc was unable to optimise the constexpr int32_t to int24_t uint32_t to
uint24_t cast which this was causing huge performance issues when
converting any type to sample<int24_t> or network_sample<int24_t>. I
solved this by creating a narrowing_cast function to handle all
narrowing casts. This function then has special cases for int32_t to
int24_t and uint32_t to uint24_t.
I have also removed the explicit usage of __builtin_memcpy as it's best
to leave this up to the compiler (which in all cases where the size is
known at compile time will optimise to use __builtin_memcpy anyway).

+++ include/ratl/detail/round.hpp
@@ -65,7 +65,7 @@ inline int32_t round_float32_to_int32_fast(float32_t input) noexcept
 #elif defined(RATL_ROUND_USE_NEON_INTRINSICS) && defined(RATL_CPP_ARCH_AARCH64)
     return vgetq_lane_s32(vcvtnq_s32_f32(vdupq_n_f32(input)), 0);
 #else
-    return static_cast<int32_t>(round_float32_to_int32_magic(input));
+    return round_float32_to_int32_magic(input);
 #endif
 }
 } // namespace detail
